feat: clean up build artifacts and update route exports for better organization
