+++ apps/react-router/app/routes/_root._index.tsx
@@ -1,3 +1,3 @@
 import RouteComponent from '@remix-docs-ja/base/routes/_root._index/route'
 export default RouteComponent
-export * from '@remix-docs-ja/base/routes/_root._index/route.server'
+export { loader } from '@remix-docs-ja/base/routes/_root._index/route.server'
