+++ apps/react-router/app/routes/_root.tsx
@@ -1,3 +1,5 @@
 import RouteComponent from '@remix-docs-ja/base/routes/_root/route'
 export default RouteComponent
-export * from '@remix-docs-ja/base/routes/_root/route.server'
+export { loader } from '@remix-docs-ja/base/routes/_root/route.server'
+
+export const shouldRevalidate = () => true
