+++ apps/react-router/app/routes/resources._index.ts
@@ -1,2 +1,1 @@
-import { loader } from '@remix-docs-ja/base/routes/resources._index'
-export { loader }
+export { loader } from '@remix-docs-ja/base/routes/resources._index'
